Moved Squadron to its own package, and added an interface
Do this so that the Squadron object can interface with the BBDLibrary.  After this commit a Squadron object IS A GameComponent also.  It can hold several GameItems and handle them accordingly, controlling how they render and interact with the user.

We made a new package to hold all the Squadron related GameItems in one centralized place.
